Updated README
I added steps at the Configuring Truenas Routing ' ' ' after step 14 detailing how to allow incoming traffic from the tailnet to be routed on the local network. Obviously changes can be made but these are the steps I had to take to get my setup 100% working and wanted to contribute as this guide got me most of the way there.